Classnote/18_AdminPanel_2/requirements.txt
asgiref==3.6.0
Django==4.1.5
django-ckeditor==6.5.1
django-js-asset==2.0.0
djangorestframework==3.14.0
Faker==15.3.4
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
